1.3.1
Made an oopsie, this time the hotfix for real

+++ PulsarPackCreator/Game Structs.cs
@@ -540,7 +540,7 @@ public class PulsarGame
         [Endian(Endianness.BigEndian)]
         public ushort cupIconCount;
         [Endian(Endianness.BigEndian)]
-        public ushort chooseNextTrackTimer;
+        public byte chooseNextTrackTimer;
         [MarshalAs(UnmanagedType.ByValArray, SizeConst = 40)]
         public byte[] reservedSpace;
     }
